Move underline to login link in VerifiedProvider component.
