+++ client/site-profiler/components/verified-provider/styles.scss
@@ -1,0 +1,3 @@
+.verified-provider {
+	color: var(--wp-admin-theme-color);
+}
